--- src/api/Api.tsx
@@ -1,7 +1,0 @@
-import { OrderItem } from '../component/OrderItem';
-
-interface Api {
-  getData(): Promise<OrderItem[]>;
-}
-
-export type { Api };
