+++ src/interfaces/Api.tsx
@@ -1,0 +1,7 @@
+import { OrderData } from './OrderItem';
+
+interface Api {
+  getData(): Promise<OrderData[]>;
+}
+
+export type { Api };
